refactor(pi-agents-tmux): the timeline renderer moves to transcript-timeline.ts at the ratchet seam
